requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile setup.py
#
asttokens==2.4.1
    # via stack-data
boto3==1.35.0
    # via example_neynar_parquet_importer (setup.py)
botocore==1.35.0
    # via
    #   boto3
    #   s3transfer
d6tcollect==1.0.6
    # via d6tstack
d6tstack==0.2.0
    # via example_neynar_parquet_importer (setup.py)
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
executing==2.0.1
    # via stack-data
ipdb==0.13.13
    # via example_neynar_parquet_importer (setup.py)
ipython==8.26.0
    # via ipdb
jedi==0.19.1
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
matplotlib-inline==0.1.7
    # via ipython
numpy==2.1.0
    # via
    #   d6tstack
    #   pandas
    #   pyarrow
    #   scipy
pandas==2.2.2
    # via
    #   d6tstack
    #   example_neynar_parquet_importer (setup.py)
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
prompt-toolkit==3.0.47
    # via ipython
psycopg2==2.9.9
    # via example_neynar_parquet_importer (setup.py)
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==17.0.0
    # via example_neynar_parquet_importer (setup.py)
pygments==2.18.0
    # via ipython
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-dotenv==1.0.1
    # via example_neynar_parquet_importer (setup.py)
pytz==2024.1
    # via pandas
s3transfer==0.10.2
    # via boto3
scipy==1.14.0
    # via d6tstack
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlalchemy==2.0.32
    # via
    #   d6tstack
    #   example_neynar_parquet_importer (setup.py)
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via sqlalchemy
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via botocore
wcwidth==0.2.13
    # via prompt-toolkit
